Add shared test helpers and command tests
