+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Test package for AtlantisBot."""
